轮播图接口完成，订单取消逻辑和商品删除逻辑未实现

+++ src/org/dao/GarouselCatalogDao.java
@@ -13,7 +13,7 @@ public interface GarouselCatalogDao {
 	/**
 	 * 维护轮播图目录
 	 */
-	public Long saveOrUpdate(GarouselCatalog gCatalog);
+	public Integer saveOrUpdate(GarouselCatalog gCatalog);
 	
 	/**
 	 * 删除轮播图目录
